Implement secure mode for DDB commands.
With these changes, certain DDB commands are marked to indicate that they
need to be secured (not allowed) when DDB secure mode is enabled.
Secure mode is controlled by the "debug.ddb.secure" sysctl (0 is disabled,
1 is enabled) and tunable. Once secure mode is enabled, it cannot be disabled
with the sysctl.

Also, a DDB_SECURE kernel configuration option is available to permanently
enable DDB secure mode and avoid compiling any of the functions that
implement the marked commands in to the kernel.